rename mirage scenarios
