+++ frontend/mirage/scenarios/unaccessible-exercise.js
