+++ frontend/mirage/scenarios/unaccessible-series.js
